S-00009 feat: added the hiring dashboard

+++ frontend/src/test/setup.js
@@ -8,6 +8,8 @@ beforeEach(() => {
 })
 
 afterEach(() => {
-  localStorage.clear()
+  if (typeof localStorage.clear === 'function') {
+    localStorage.clear()
+  }
   vi.restoreAllMocks()
 })
